Fix problem in cal mapping at end of line
If you move left twice when you're at the end of the line with `l` in
'whichwrap', then when we search backwards we can find a digit/sum at
the start of the following line. Not sure if it's possible to fix by
changing the regexp (failed after five mins of messing around with
`/\_`) so instead brute-forced it by temporarily removing `l` from
'whichwrap'.