Add VimrcHTML pattern
Attempt to fix layout on small screens with responsive design.

Can't remember the details because most of it I wrote a very long time
ago but never commited because it's incomplete. (Most notably, while
almost all the content is rendered within the viewport, the page is
still mysteriously almost twice as it needs to be so you can scroll
right to view a big empty space.)